rename models dir to model dir
