+++ lib/myob-api.rb
@@ -3,7 +3,7 @@ require 'myob/api/version'
 require 'myob/api/helpers'
 require 'myob/api/query'
 
-require 'myob/api/models/base'
-require 'myob/api/models/contact'
+require 'myob/api/model/base'
+require 'myob/api/model/contact'
 
 require 'myob/api/client'
