+++ lib/myob/api/model/contact.rb
